Embed the Grammarly server.js to coc-vim
This is the easiest way I figured out to import the Grammarly project.
Unfortunately it is not easy to import from npm because it depends from
something in vscode that I can't figure out.

There is an issue open to split the lsp from the vscode engine. Until
then I think we do not have other options.